requirements.txt
pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu118
transformers==4.56.2
accelerate==0.28.0
bitsandbytes==0.43.1
huggingface-hub==0.22.2
langchain==0.1.14
langchain-core==0.1.43
langchain-community==0.0.31
pypdf==4.2.0
sentence-transformers==5.1.0
beautifulsoup4==4.12.3
langserve[all]
faiss-cpu
rapidocr-onnxruntime==1.4.1
unstructured==0.18.5
fastapi==0.110.1
uvicorn==0.29.0
wget
rank_bm25